Add jQuery, jQuery-UJS, Brunch, etc
This looks icky, but jQuery/UJS are how we can use proper HTTP verbs
with our endpoints. It also gets us out of trusting someone else to host
our CSS/JS for us (what we deploy is what we serve).

This does mean we take a dependency on NodeJS for development, which I'm
not happy about, but is a fair tradeoff for what it delivers.